+++ web/app/dashboard/page.tsx
@@ -9,6 +9,7 @@ import { GitHubOperationalWorkPanel } from "../../components/GitHubOperationalWo
 import { GitHubSyncControls } from "../../components/GitHubSyncControls";
 import { PageHeader } from "../../components/PageHeader";
 import { PrivateBetaReadinessPanel } from "../../components/PrivateBetaReadinessPanel";
+import { RepositoryAuditOverviewPanel } from "../../components/RepositoryAuditOverviewPanel";
 import { SelectedRepositorySyncControls } from "../../components/SelectedRepositorySyncControls";
 import { SourceCoveragePanel } from "../../components/SourceCoveragePanel";
 import { StatusCard } from "../../components/StatusCard";
@@ -62,6 +63,7 @@ export default function DashboardPage() {
       />
       <SourceCoveragePanel refreshSignal={operationalWorkRefresh} />
       <PrivateBetaReadinessPanel refreshSignal={operationalWorkRefresh} />
+      <RepositoryAuditOverviewPanel refreshSignal={operationalWorkRefresh} />
       <CompanyBrainPanel refreshSignal={operationalWorkRefresh} />
       <BriefingPanel />
       <ActionProposalsPanel />